Tiles, Transparency, Parallax, Destroyable Grass
Added four complete rule tile sets for 'drawing' levels along with layers for the various types of objects for the environment.

Enabled a transparency effect that lets allows the player character to be visible behind trees and the canopy.

Added a parallax effect to the canopy.

Created destroyable grass that creates a particle effect when hit by the sword.