Inventory and New Weapons Pt. 1
Created an inventory system with three item slots that allows us to toggle between active weapons.

Created an interface class that allows us to call a generic attack function, which will then use whatever attack is associated with our active weapon.

Created scriptable objects for the bow and staff in preparation for adding two new weapons to the game.