Shooter Class + First New Enemy
Created a shooter class that handles a ton of logic for enemies to shoot bullets at the player. This includes firing multiple shots in burst, at intervals, in arcs, and oscillating, all while constantly shooting at the player.

Created a new Ghost enemy as the first enemy to have this shooter class.